# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/payment/calculate-structure": {
      "post": {
        "operationId": "PaymentsController_calculateStructure",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Calculate the full payment structure",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payment/calculate-deposit": {
      "post": {
        "operationId": "PaymentsController_calculateDeposit",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Calculate the deposit and platform fee",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payment/calculate-installments": {
      "post": {
        "operationId": "PaymentsController_calculateInstallments",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Calculate the installment schedule",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payment/update-balance": {
      "post": {
        "operationId": "PaymentsController_updateBalance",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Compute the remaining balance",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payment/update-status": {
      "post": {
        "operationId": "PaymentsController_updateStatus",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Compute the next payment status",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/transactions": {
      "get": {
        "operationId": "TransactionsController_getTransactions",
        "parameters": [
          {
            "name": "includeReceiptSignedUrls",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "receiptType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the current user's transaction history (paginated)",
        "tags": [
          "transactions"
        ]
      }
    },
...
    "/api/v1/admin/onboard-school": {
      "post": {
        "operationId": "AdminController_onboardSchool",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchoolDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Onboard a new school",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/paystack/banks": {
      "get": {
        "operationId": "AdminController_getBanks",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List Nigerian banks for the settlement-bank dropdown",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/paystack/resolve-account": {
      "post": {
        "operationId": "AdminController_resolveAccount",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resolve an account number + bank code to the registered account name",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/schools/{schoolId}/paystack-subaccount": {
      "post": {
        "operationId": "AdminController_createSubaccount",
        "parameters": [
          {
            "name": "schoolId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a Paystack subaccount for a school missing one",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/pending-first-payments": {
      "get": {
        "operationId": "AdminController_getPendingFirstPayments",
        "parameters": [
          {
            "name": "includeReceiptSignedUrls",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List pending first payments (paginated)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/pending-installments": {
      "get": {
        "operationId": "AdminController_getPendingInstallments",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List pending installment payments across all schools (paginated)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/schools/{schoolId}/students": {
      "get": {
        "operationId": "AdminController_getSchoolStudents",
        "parameters": [
          {
            "name": "schoolId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "className",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List students/enrollments for a specific school (paginated)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/settle-first-payment/{paymentId}": {
      "post": {
        "operationId": "AdminController_settleFirstPayment",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Settle a first payment (release the school share)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/reject-first-payment/{paymentId}": {
      "post": {
        "operationId": "AdminController_rejectFirstPayment",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject a first payment",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/revenue": {
      "get": {
        "operationId": "AdminController_getRevenue",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get platform revenue",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/transactions": {
      "get": {
        "operationId": "AdminController_getTransactions",
        "parameters": [
          {
            "name": "includeReceiptSignedUrls",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "receiptType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List global transactions across all schools (paginated)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/students/summary": {
      "get": {
        "operationId": "AdminController_getStudentsSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get global student summary",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/schools/summary": {
      "get": {
        "operationId": "AdminController_getSchoolsSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get per-school summary",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/overview": {
      "get": {
        "operationId": "AdminController_getOverview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get admin overview (single-call dashboard payload)",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "LoPay API",
    "description": "The LoPay API documentation for frontend integration",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSchoolDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}